A start
Frankly, there's no use for this plugin. But someone wanted it and I was curious to see what could be tinkered with. The answer? Almost nothing. None of this junk needs to be exposed to blueprints. Maybe I'll repurpose this space to write some custom functions which actually are useful. Probably not. 99% chance this just becomes a waste of storage space.